Upgrade lbeta arguments from ints to doubles

+++ cxx/util_math.cc
@@ -8,7 +8,9 @@
 #include <random>
 
 // http://matlab.izmiran.ru/help/techdoc/ref/betaln.html
-double lbeta(int z, int w) { return lgamma(z) + lgamma(w) - lgamma(z + w); }
+double lbeta(double z, double w) {
+  return lgamma(z) + lgamma(w) - lgamma(z + w);
+}
 
 std::vector<double> linspace(double start, double stop, int num,
                              bool endpoint) {
